Add json data to assets
During development json data will not be downloaded
bu loading from a bundled file.
To not add this file to a release version it
is added to debug. This means that release version
will not work properly unless the http
fetching is implemented or the local json file
is copied to the release assets.